Extract notifier from task spec
