+++ collect_app/src/main/java/org/odk/collect/android/notifications/Notifier.java
@@ -1,0 +1,12 @@
+package org.odk.collect.android.notifications;
+
+import org.odk.collect.android.formmanagement.ServerFormDetails;
+
+import java.util.HashMap;
+
+public interface Notifier {
+
+    void onUpdatesAvailable();
+
+    void onUpdatesDownloaded(HashMap<ServerFormDetails, String> result);
+}
